json parser returns int as float64

+++ main.go
@@ -19,12 +19,12 @@ func main() {
 	}
 	log.Println("Starting modbus..")
 	port, ok1 := settings.Value("Port", "/dev/ttyS0").(string)
-	baudRate, ok2 := settings.Value("BoudRate", 57600).(int)
+	baudRate, ok2 := settings.Value("BoudRate", 57600).(float64)
 	rtsPin, ok3 := settings.Value("RtsPin", "gpio3_pg8").(string)
 	if ok1 && ok2 && ok3 {
 		dataToServerChan = make(chan cubiectrl.CellData)
 		
-		resChan, err := StartModbusClient(port, baudRate, rtsPin, settings)
+		resChan, err := StartModbusClient(port, int(baudRate), rtsPin, settings)
 		if err != nil {
 			log.Fatal(err)
 		} else {
